+++ src/frydom/hydrodynamics/manoeuvring/FrAbkowitzManoeuvringForce.cpp
@@ -3,20 +3,21 @@
 //
 
 #include "FrAbkowitzManoeuvringForce.h"
+#include "frydom/environment/FrEnvironmentInc.h"
 #include "frydom/io/JSONNode.h"
 
 namespace frydom {
 
   FrAbkowitzManoeuvringForce::FrAbkowitzManoeuvringForce(const std::string &name,
                                                          FrBody *body,
-                                                         const std::string &file) :
+                                                         const std::string &file,
+                                                         const Position& mid_ship) :
       FrForce(name, "FrManoeuvringForce", body),
       c_filepath(file),
-      m_Xvv(0.), m_Xvvvv(0.), m_Xvvu(0.), m_Xrr(0.), m_Xrru(0.), m_Xvr(0.), m_Xvru(0.),
-      m_Yv(0.), m_Yvvv(0.), m_Yvrr(0.), m_Yvu(0.), m_Yvuu(0.), m_Yr(0.), m_Yrrr(0.), m_Yvvr(0.), m_Yru(0.), m_Yruu(0.),
-      m_Nv(0.), m_Nvvv(0.), m_Nvrr(0.), m_Nvu(0.), m_Nvuu(0.), m_Nr(0.), m_Nrrr(0.), m_Nvvr(0.), m_Nru(0.), m_Nruu(0.) {
-
-  }
+      m_Xvv(0.), m_Xvvvv(0.), m_Xrr(0.), m_Xvr(0.),
+      m_Yv(0.), m_Yvvv(0.), m_Yvrr(0.), m_Yr(0.), m_Yrrr(0.), m_Yvvr(0.),
+      m_Nv(0.), m_Nvvv(0.), m_Nvrr(0.), m_Nr(0.), m_Nrrr(0.), m_Nvvr(0.),
+      m_mid_ship(mid_ship) {}
 
   void FrAbkowitzManoeuvringForce::Initialize() {
     FrForce::Initialize();
@@ -24,6 +25,61 @@ namespace frydom {
   }
 
   void FrAbkowitzManoeuvringForce::Compute(double time) {
+
+    auto body = GetBody();
+    auto rho = GetSystem()->GetEnvironment()->GetFluidDensity(WATER);
+
+    // Get the velocity at mid-ship
+    Velocity vessel_vel_world = body->GetVelocityInWorldAtPointInBody(m_mid_ship, NWU);
+    auto vessel_angular_vel_world = body->GetAngularVelocityInWorld(NWU);
+
+    // Projection to the planar frame
+    Direction vessel_x_axis_world = body->GetFrame().GetXAxisInParent(NWU);
+    vessel_x_axis_world[2] = 0;
+    vessel_x_axis_world.Normalize();
+
+    Direction vessel_y_axis_world = body->GetFrame().GetYAxisInParent(NWU);
+    vessel_y_axis_world[2] = 0;
+    vessel_y_axis_world.Normalize();
+
+    double u = vessel_vel_world.dot(vessel_x_axis_world);
+    double v = vessel_vel_world.dot(vessel_y_axis_world);
+    double r = vessel_angular_vel_world.GetWz();
+
+    // Adim
+    double vnorm2 = u*u + v*v;
+    double vnorm = std::sqrt(vnorm2);
+    auto q = 0.5 *rho * m_Lpp * m_draft * vnorm2;
+
+    // Compute hydrodynamic derivatives
+    double c_Xvv = m_Xvv * v*v;
+    double c_Xvr = m_Xvr * v*r;
+    double c_Xrr = m_Xrr * r*r;
+    double c_Xvvvv = m_Xvvvv * pow(v, 4);
+
+    double c_Yv = m_Yv * v;
+    double c_Yr = m_Yr * r;
+    double c_Yvvv = m_Yvvv * pow(v, 3);
+    double c_Yvvr = m_Yvvr * v*v*r;
+    double c_Yvrr = m_Yvrr * v*r*r;
+    double c_Yrrr = m_Yrrr * r*r*r;
+
+    double c_Nv = m_Nv * v;
+    double c_Nr = m_Nr * r;
+    double c_Nvvv = m_Nvvv * pow(v, 3);
+    double c_Nvvr = m_Nvvr * v*v*r;
+    double c_Nvrr = m_Nvrr * v*r*r;
+    double c_Nrrr = m_Nrrr * pow(r, 3);
+
+    // Compute force and torque
+    auto X = q * (c_Xvv + c_Xvr + c_Xrr + c_Xvvvv);
+    auto Y = q * (c_Yv + c_Yr + c_Yvvv + c_Yvvr + c_Yvrr + c_Yrrr);
+    auto N = q * m_Lpp * (c_Nv + c_Nr + c_Nvvv + c_Nvvr + c_Nvrr + c_Nrrr);
+
+    Force forceInWorld = X * vessel_x_axis_world + Y * vessel_y_axis_world;
+    Torque torqueInWorld = {0., 0., N};
+
+    SetForceTorqueInWorldAtCOG(forceInWorld, torqueInWorld, NWU);
 
   }
 
@@ -58,31 +114,28 @@ namespace frydom {
 
     m_Xvv = coeffs_node.get_val<double>("Xvv", 0.0);
     m_Xvvvv = coeffs_node.get_val<double>("Xvvvv", 0.0);
-    m_Xvvu = coeffs_node.get_val<double>("Xvvu", 0.0);
     m_Xrr = coeffs_node.get_val<double>("Xrr", 0.0);
-    m_Xrru = coeffs_node.get_val<double>("Xrru", 0.0);
     m_Xvr = coeffs_node.get_val<double>("Xvr", 0.0);
-    m_Xvru = coeffs_node.get_val<double>("Xvru", 0.0);
     m_Yv = coeffs_node.get_val<double>("Yv", 0.0);
     m_Yvvv = coeffs_node.get_val<double>("Yvvv", 0.0);
     m_Yvrr = coeffs_node.get_val<double>("Yvrr", 0.0);
-    m_Yvu = coeffs_node.get_val<double>("Yvu", 0.0);
-    m_Yvuu = coeffs_node.get_val<double>("Yvuu", 0.0);
     m_Yr = coeffs_node.get_val<double>("Yr", 0.0);
     m_Yrrr = coeffs_node.get_val<double>("Yrrr", 0.0);
     m_Yvvr = coeffs_node.get_val<double>("Yvvr", 0.0);
-    m_Yru = coeffs_node.get_val<double>("Yru", 0.0);
-    m_Yruu = coeffs_node.get_val<double>("Yruu", 0.0);
     m_Nv = coeffs_node.get_val<double>("Nv", 0.0);
     m_Nvvv = coeffs_node.get_val<double>("Nvvv", 0.0);
     m_Nvrr = coeffs_node.get_val<double>("Nvrr", 0.0);
-    m_Nvu = coeffs_node.get_val<double>("Nvu", 0.0);
-    m_Nvuu = coeffs_node.get_val<double>("Nvuu", 0.0);
     m_Nr = coeffs_node.get_val<double>("Nr", 0.0);
     m_Nrrr = coeffs_node.get_val<double>("Nrrr", 0.0);
     m_Nvvr = coeffs_node.get_val<double>("Nvvr", 0.0);
-    m_Nru = coeffs_node.get_val<double>("Nru", 0.0);
-    m_Nruu = coeffs_node.get_val<double>("Nruu", 0.0);
+  }
+
+  std::shared_ptr<FrAbkowitzManoeuvringForce>
+      make_abkowitz_manoeuvring_model(const std::string& name, std::shared_ptr<FrBody> body,
+                                      const std::string& file, const Position& mid_ship) {
+    auto force = std::make_shared<FrAbkowitzManoeuvringForce>(name, body.get(), file, mid_ship);
+    body->AddExternalForce(force);
+    return force;
   }
 
 }  // end namespace frydom